fix: harden dependency audit and build workflow
