+++ frontend/src/components/WikiCard.tsx
@@ -1,7 +1,7 @@
 import { Heart, Share2 } from "lucide-react"
 import { useState } from "react"
-import { useLikedArticles } from "../contexts/LikedArticlesContext"
-import { useToast } from "../contexts/ToastContext"
+import { useLikedArticles } from "../contexts/likedArticles"
+import { useToast } from "../contexts/toast"
 import { useI18n } from "../hooks/useI18n"
 import "../styles/WikiCard.css"
 import type { ArticleProps } from "../types/ArticleProps"
