+++ frontend/src/contexts/toast.ts
@@ -1,0 +1,19 @@
+import { createContext, useContext } from 'react';
+
+export interface ToastOptions {
+  duration?: number;
+}
+
+export interface ToastContextValue {
+  showToast: (message: string, options?: ToastOptions) => void;
+}
+
+export const ToastContext = createContext<ToastContextValue | undefined>(undefined);
+
+export function useToast(): ToastContextValue {
+  const ctx = useContext(ToastContext);
+  if (!ctx) {
+    throw new Error('useToast must be used within a ToastProvider');
+  }
+  return ctx;
+}
